fib 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,15 +7,15 @@ int main(void)
 {
 	int n = 98;
 	int i;
-	int first_term = 1;
-	int second_term = 2;
-	int next_term ;
-	printf("%d,%d", first_term, second_term);
+	long int first_term = 1;
+	long int second_term = 2;
+	long int next_term ;
+	printf("%lu,%lu", first_term, second_term);
 
 	for (i = 2 ; i < n; i++)
 	{
 		next_term = first_term + second_term;
-		printf("%d", next_term);
+		printf("%lu", next_term);
 		if (!(i == (n - 1)))
 		{
 			printf(", ");
